fix(ci): python-coverage-shards — split oversized part_006 to satisfy first-party line limit
